17th_commit

+++ view.php
@@ -42,9 +42,10 @@
             <img src="<?= $animals['image']; ?>" width=200 height=200>
             <P> Posted By: <?= $animals['user']; ?> </P>
             <p> Posted on: <?= $animals['timestamp']; ?> </p>
-            <span><a href="edit2.php?id=<?= $animals['id']; ?>">Edit</a></span>
-            <span><a href="delete2.php?id=<?= $animals['id']; ?>">Delete</a></span>
         <?php endforeach; ?>
+
+        <span><a href="edit2.php?id=<?= $animals['id']; ?>">Edit</a></span>
+            <span><a href="delete2.php?id=<?= $animals['id']; ?>" onclick="return confirm('Are you sure?')">Delete</a></span>
 
         <h2> Your Comments: </h2>
         <?php foreach ($p as $comments): ?>
